add components, fixed
