+++ src/features/index.ts
@@ -2,3 +2,5 @@ export { default as TypeFilmList } from './TypeFilmList/ui';
 export { default as BackButton } from './BackButton/ui';
 export { default as MoviesSearch } from './MoviesSearch/ui';
 export { default as SelectViewButtons } from './SelectViewButtons/ui';
+export { default as TitleInfoPage } from './TitleInfoPage/ui';
+export { default as ActorsSearch } from './ActorsSearch/ui';
