+++ src/pages/ActorInfoPage/ui/index.tsx
@@ -6,13 +6,12 @@ import {
   useParams,
 } from 'react-router-dom';
 import styled from 'styled-components';
-import { format } from 'date-fns';
-import ruLocale from 'date-fns/locale/ru';
 import { BiChevronLeft, BiChevronsDown } from 'react-icons/bi';
 import baseTheme, { textGradient } from '../../../shared/styles/theme';
 import { useRootStore } from '../../../shared/libs/hooks';
 import { Card, Spinner } from '../../../shared/ui';
 import { TActorInfo } from '../../../entities';
+import { ActorInfo } from '../../../widgets';
 
 const SpinnerContainer = styled.div`
   width: 100%;
@@ -60,52 +59,9 @@ const Poster = styled.img`
   margin-bottom: 1rem;
 `;
 
-const InfoContainer = styled.div<{poster: boolean}>`
-  width: ${props => (props.poster ? 'calc(100% - 40vh - 1rem)' : '100%')};
-  display: flex;
-  flex-direction: column;
-  row-gap: 0.35rem;
-  height: max-content;
-  @media ${baseTheme.media.l} {
-    width: ${props => (props.poster ? 'calc(100% - 35vh - 1rem)' : '100%')};
-  }
-  @media ${baseTheme.media.m} {
-    width: 100%;
-    align-items: center;
-  }
-`;
-
-const MainTitle = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 1rem;
-  @media ${baseTheme.media.m} {
-    align-items: center;
-  }
-`;
-
-const ActorsContainer = styled.div`
+const MoviesContainer = styled.div`
   width: 100%;
   margin-top: 2.5rem;
-`;
-
-const Title = styled.div`
-  font-size: 3rem;
-  font-weight: 400;
-  line-height: 3.2rem;
-  @media ${baseTheme.media.l} {
-    font-size: 2rem;
-    line-height: 2.2rem;
-  }
-`;
-
-const Description = styled.div`
-  font-weight: 400;
-  display: flex;
-  align-items: center;
-  @media ${baseTheme.media.m} {
-    text-align: center;
-  }
 `;
 
 const SubTitle = styled.div`
@@ -240,95 +196,15 @@ const ActorInfoPage: React.FC = () => {
               alt=""
             />
           </PosterContainer>
-          <InfoContainer poster={!!actorInfoStore.actorInfo.photo}>
-            <MainTitle>
-              <Title>
-                {actorInfoStore.actorInfo.name}
-              </Title>
-              {
-                  actorInfoStore.actorInfo.enName && <Title>({actorInfoStore.actorInfo.enName})</Title>
-                }
-            </MainTitle>
-            {
-              actorInfoStore.actorInfo.sex && (
-                <Description>
-                  Пол: {actorInfoStore.actorInfo.sex}
-                </Description>
-              )
-            }
-            {
-              actorInfoStore.actorInfo.age && (
-                <Description>
-                  Возраст: {actorInfoStore.actorInfo.age}
-                </Description>
-              )
-            }
-            {
-              actorInfoStore.actorInfo.birthday && (
-                <Description>
-                  Дата рождения: {format(new Date(actorInfoStore.actorInfo.birthday), 'dd MMMM Y', { locale: ruLocale })}
-                </Description>
-              )
-            }
-            {
-              actorInfoStore.actorInfo.birthPlace?.length && (
-                <Description>
-                  Место рождения: {actorInfoStore.actorInfo.birthPlace?.map((i, index, arr) => {
-                  const comma = index === arr.length - 1 ? '' : ', ';
-                  return i.value + comma;
-                })}
-                </Description>
-              )
-            }
-            {
-              actorInfoStore.actorInfo.growth && (
-                <Description>
-                  Рост: {actorInfoStore.actorInfo.growth}
-                </Description>
-              )
-            }
-            {
-              actorInfoStore.actorInfo.countAwards && (
-                <Description>
-                  Количество наград: {actorInfoStore.actorInfo.countAwards}
-                </Description>
-              )
-            }
-            <Description>
-              Карьера: {actorInfoStore.actorInfo.profession?.map((i, index, arr) => {
-              const comma = index === arr.length - 1 ? '' : ', ';
-              return i.value + comma;
-            })}
-            </Description>
-            <Description>
-              Количество фильмов: {actorInfoStore.actorInfo.movies?.length ?? 0}
-            </Description>
-            {
-              actorInfoStore.actorInfo?.death && (
-                <Description>
-                  Дата смерти: {format(new Date(actorInfoStore.actorInfo.death), 'dd MMMM Y', { locale: ruLocale })}
-                </Description>
-              )
-            }
-            {
-              !!actorInfoStore.actorInfo.deathPlace?.length && (
-                <Description>
-                  Место смерти: {actorInfoStore.actorInfo.deathPlace?.map((i, index, arr) => {
-                  const comma = index === arr.length - 1 ? '' : ', ';
-                  return i.value + comma;
-                })}
-                </Description>
-              )
-            }
-          </InfoContainer>
-          <ActorsContainer>
+          <ActorInfo actor={actorInfoStore.actorInfo} />
+          <MoviesContainer>
             <SubTitle>
               Фильмы c {actorInfoStore.actorInfo.sex === 'Мужской' ? 'его' : 'её'} участием
             </SubTitle>
             <Movies>
               {
                 filmList?.map(i => (
-                  <MovieCard to={`/movie/${i.id}`}>
+                  <MovieCard key={i.id} to={`/movie/${i.id}`}>
                     <Card>
                       <MovieContent>
                         <MovieTitle>{i.name}</MovieTitle>
@@ -359,7 +235,7 @@ const ActorInfoPage: React.FC = () => {
                 }
               </div>
             </Movies>
-          </ActorsContainer>
+          </MoviesContainer>
         </ContentContainer>
         )
       }
