+++ src/pages/PremieresPage/ui/index.tsx
@@ -2,29 +2,18 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { observer } from 'mobx-react-lite';
 import routes from '../../../shared/routes';
-import { PageContainer, Title, Select } from '../../../shared/ui';
+import { PageContainer, Title } from '../../../shared/ui';
 import { useRootStore } from '../../../shared/libs/hooks';
 import {
-  IOption, MovieList, months, years,
+  MovieList, SelectPremierDate, months, years,
 } from '../../../widgets';
 import { TPremiereFilters } from '../../../entities';
 import { SelectViewButtons } from '../../../features';
+import { IOption } from '../../../shared/types';
 
 const ModeButtons = styled.div`
   margin-left: auto;
   padding-top: 1rem;
-`;
-
-const FiltersContainer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  margin-top: 1rem;
-  margin-bottom: 1.4rem;
-`;
-
-const Filter = styled.div`
-  width: 49%;
 `;
 
 const PremieresPage: React.FC = () => {
@@ -44,22 +33,12 @@ const PremieresPage: React.FC = () => {
           <SelectViewButtons />
         </ModeButtons>
       </div>
-      <FiltersContainer>
-        <Filter>
-          <Select
-            options={months}
-            selected={month}
-            setSelected={setMonth}
-          />
-        </Filter>
-        <Filter>
-          <Select
-            options={years}
-            selected={year}
-            setSelected={setYear}
-          />
-        </Filter>
-      </FiltersContainer>
+      <SelectPremierDate
+        year={year}
+        month={month}
+        setMonth={setMonth}
+        setYear={setYear}
+      />
       <MovieList
         type="PREMIERES"
         premiereFilters={{
