+++ src/shared/ui/SubTitle.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const SubTitle = styled.div`
+  font-size: 1.75rem;
+  margin-bottom: 0.5rem;
+  display: flex;
+  align-items: center;
+`;
+
+export default SubTitle;
