+++ src/widgets/SimilarMovies/types/similarMovie.ts
@@ -1,0 +1,3 @@
+import { TMovieInfo } from '../../../entities';
+
+export type TSimilarMovies = TMovieInfo['similarMovies'];
