+++ src/widgets/index.ts
@@ -1,12 +1,12 @@
-export { default as ActorList } from './ActorList/ui/index';
-export { default as CountryList } from './CountryList/ui/index';
-export { default as GenreList } from './GenreList/ui/index';
-export { default as Header } from './Header/ui/index';
-export { default as MovieList } from './MovieList/ui/index';
-export { default as SelectPremierDate } from './SelectPremierDate/ui/index';
-export { default as YearList } from './YearList/ui/index';
-
-export type { IOption } from './SelectPremierDate/types/options';
+export { default as ActorList } from './ActorList/ui';
+export { default as CountryList } from './CountryList/ui';
+export { default as GenreList } from './GenreList/ui';
+export { default as Header } from './Header/ui';
+export { default as MovieList } from './MovieList/ui';
+export { default as SelectPremierDate } from './SelectPremierDate/ui';
+export { default as YearList } from './YearList/ui';
+export { default as ActorInfo } from './ActorInfo/ui';
+export { default as SimilarMovies } from './SimilarMovies/ui';
 
 export { default as years } from './SelectPremierDate/consts/years';
 export { default as months } from './SelectPremierDate/consts/months';
